+++ src/Actions/ManagesTrees.php
@@ -2,6 +2,7 @@
 
 namespace Rickkuilman\DigitalHumaniPhpSdk\Actions;
 
+use Exception;
 use Rickkuilman\DigitalHumaniPhpSdk\Resources\Tree;
 
 trait ManagesTrees
@@ -9,22 +10,23 @@ trait ManagesTrees
     /**
      * Plant one or many trees
      *
-     * @param string $enterpriseId Id of your enterprise. Example of an enterprise id: 11111111 (Enterprise
-     * Ids are 8 digits long)
      * @param string $user End user by whom the trees were planted. Example of an user: email@test.com
      * @param int $amount Number of trees requested to plant. Example: 1
      * @param string|null $projectId Id of the reforestation project for where you want the trees to be planted.
      * Example of an id: 93333333 (Project Ids are 8 digits long)
+     * @param string|null $enterpriseId Id of your enterprise. Example of an enterprise id: 11111111 (Enterprise
+     * Ids are 8 digits long)
      * @return Tree
+     * @throws Exception
      */
-    public function plantTree(string $enterpriseId, string $user, int $amount = 1, string $projectId = null): Tree
+    public function plantTree(string $user, int $amount = 1, string $projectId = null, string $enterpriseId = null): Tree
     {
         if (is_null($projectId)) {
             $projectId = $this->defaultProjectId();
         }
 
         return new Tree($this->post('tree', [
-            'enterpriseId' => $enterpriseId,
+            'enterpriseId' => $enterpriseId ?? $this->getDefaultEnterpriseId(),
             'user' => $user,
             'treeCount' => $amount,
             'projectId' => (string)$projectId,
@@ -36,6 +38,7 @@ trait ManagesTrees
      *
      * @param string $uuid uuid of the trees for which you want to get the details.
      * @return Tree
+     * @throws Exception
      */
     public function tree(string $uuid): Tree
     {
@@ -45,15 +48,16 @@ trait ManagesTrees
     /**
      * Count trees for a user
      *
-     * @param string $enterpriseId Id of your enterprise. Example of an enterprise id: 11111111 (Enterprise
-     * Ids are 8 digits long)
      * @param string $user End user by whom the trees were planted. Example of an user: email@test.com
+     * @param string|null $enterpriseId Id of your enterprise. Example of an enterprise id: 11111111 (Enterprise
+     * Ids are 8 digits long)
      * @return array
+     * @throws Exception
      */
-    public function countTreesPlantedByUser(string $enterpriseId, string $user): array
+    public function countTreesPlantedByUser(string $user, string $enterpriseId = null): array
     {
         return $this->get(sprintf("tree?enterpriseId=%s&user=%s",
-            $enterpriseId,
+            $enterpriseId ?? $this->getDefaultEnterpriseId(),
             $user
         ));
     }