+++ src/Exceptions/BadRequestException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rickkuilman\DigitalHumaniPhpSdk\Exceptions;
+
+use Exception;
+
+class BadRequestException extends Exception
+{
+
+}