+++ tests/DigitalHumaniPhpSdkTest.php
@@ -7,6 +7,7 @@ use GuzzleHttp\Psr7\Response;
 use Mockery;
 use PHPUnit\Framework\TestCase;
 use Rickkuilman\DigitalHumaniPhpSdk\DigitalHumani;
+use Rickkuilman\DigitalHumaniPhpSdk\Exceptions\BadRequestException;
 use Rickkuilman\DigitalHumaniPhpSdk\Exceptions\ForbiddenException;
 use Rickkuilman\DigitalHumaniPhpSdk\Exceptions\NotFoundException;
 use Rickkuilman\DigitalHumaniPhpSdk\Exceptions\UnauthorizedException;
@@ -20,7 +21,7 @@ class DigitalHumaniPhpSdkTest extends TestCase
 
     public function test_making_basic_requests()
     {
-        $digitalHumani = new DigitalHumani('abc', false, $http = Mockery::mock(Client::class));
+        $digitalHumani = new DigitalHumani('abc', '4c6e672d', false, $http = Mockery::mock(Client::class));
 
         $http->shouldReceive('request')
             ->once()
@@ -36,7 +37,7 @@ class DigitalHumaniPhpSdkTest extends TestCase
     {
         $this->expectException(NotFoundException::class);
 
-        $digitalHumani = new DigitalHumani('abc', false, $http = Mockery::mock(Client::class));
+        $digitalHumani = new DigitalHumani('abc', '4c6e672d', false, $http = Mockery::mock(Client::class));
 
         $http->shouldReceive('request')
             ->once()
@@ -52,7 +53,7 @@ class DigitalHumaniPhpSdkTest extends TestCase
     {
         $this->expectException(UnauthorizedException::class);
 
-        $digitalHumani = new DigitalHumani('abc', false, $http = Mockery::mock(Client::class));
+        $digitalHumani = new DigitalHumani('abc', '4c6e672d', false, $http = Mockery::mock(Client::class));
 
         $http->shouldReceive('request')
             ->once()
@@ -68,7 +69,7 @@ class DigitalHumaniPhpSdkTest extends TestCase
     {
         $this->expectException(ForbiddenException::class);
 
-        $digitalHumani = new DigitalHumani('abc', false, $http = Mockery::mock(Client::class));
+        $digitalHumani = new DigitalHumani('abc', '4c6e672d', false, $http = Mockery::mock(Client::class));
 
         $http->shouldReceive('request')
             ->once()
@@ -78,6 +79,29 @@ class DigitalHumaniPhpSdkTest extends TestCase
             );
 
         $digitalHumani->projects();
+    }
+
+    public function test_handling_400_errors()
+    {
+        $this->expectException(BadRequestException::class);
+
+        $digitalHumani = new DigitalHumani('abc', '4c6e672d', false, $http = Mockery::mock(Client::class));
+
+        $http->shouldReceive('request')
+            ->once()
+            ->with('POST', 'tree', [
+                'json' => [
+                    'enterpriseId' => '4c6e672d',
+                    'user' => 'rick@example.com',
+                    'treeCount' => 1,
+                    'projectId' => DigitalHumani::DEFAULT_PROJECT_ID,
+                ],
+            ])
+            ->andReturn(
+                new Response(400)
+            );
+
+        $digitalHumani->plantTree('rick@example.com');
     }
 
 }