+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 All notable changes to `digital-humani-php-sdk` will be documented in this file
 
-## 1.0.0 - 2021-08-08
+## 2.0.0 - 2021-10-23
 
-- Initial release 🎉
+- Added ability to set a default enterprise for all requests
+- Added ability to plant a tree right from an Enterprise
 
 ## 1.1.0 - 2021-10-05
 
 - Refactored package to plain PHP package (removed Laravel specific files)
 - Added ability to switch between sandbox and production environment
 - Renamed package (and namespace) to emphasize the fact that this is a "PHP" SDK
+
+
+## 1.0.0 - 2021-08-08
+
+- Initial release 🎉